not display message fix

+++ timeOrganizer.py
@@ -13,11 +13,11 @@ while True:
         m=s/60
         s=s%60
         if m>60:
-            h=m/60
+            h=int(m/60)
             m=m%60
 
     print("%s:%s:%s" %( int(h),int(m),int(s)))
 
-    if h %3==2:
+    if int(h) %3==2:
         print("eleg a kockulasbol TANULJ, PROGRAMOZZ CSINALJ VALAMI HASZNOSAT IS !!!")
